chore: move all the tests in mobile to test/mobile
