+++ test/mobile/toolbar/mobile/mobile_toolbar_test.dart
@@ -1,7 +1,7 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import '../../infra/testable_editor.dart';
+import '../../../new/infra/testable_editor.dart';
 import 'test_helpers/mobile_app_with_toolbar_widget.dart';
 
 void main() {
