+++ test/mobile/toolbar/mobile/utils/mobile_toolbar_item_menu_btn_test.dart
@@ -15,6 +15,7 @@ void main() {
           onPressed: () {},
           icon: icon,
           label: label,
+          isSelected: false,
         ),
       ),
     );
